This submission contains way too  many changes, going back to mid-September.  I waited to turn them in while trying get everything I wanted to accomplish to work correctly.  It's not at that point yet, but the changes really need to be submitted.
The main thrust of the submission is adding an algorithm to find arcs in the set of super points.  The hope is that the arc finding, in conjunction with the existing line finding, will allow feature level objects to be joined to form pipe networks.  The algorithm does find arcs, but misses some and finds some that it shouldn't.  While working on the arc finding, numerous other changes were made.
* The ChainFinder.py file was renamed to FeatureFinder.py, and that is where the serious algorithmic stuff lives.
* The line finding code in that file was improved over that in the previous submission.
* Separate test files were created for finding lines, arcs, and both lines and arcs.
* Code was added to produce 2D debugging graphs of points, lines and acs.
* The code for 3D debugging displays was cleaned up and made more useful.
* Assorted code refactoring and clean-up was performed, but there are still many "TODOs" to be found.

A clustering step was introduced in Clustering.py.  It uses KD trees group adjacent points that have similar radii into a "clustered" point.  It is these points that are now used to find lines and arcs.  This speeds things up and simplifies that algorithms because they don't have to deal with a bunch of nearly coincident points.

The current algorithms for finding arcs and finding both lines and arcs still need more work.  It would be best to do that with a new generation of super point data that is generated from the newest ML model and at a finer spatial resolution.  But the algorithms produce enough data right now that work can begin on joining lines and arcs to form pipe networks.